requirements.txt
streamlit
openai
chromadb
python-dotenv
sentence-transformers
pandas
matplotlib
seaborn
numpy